Development

+++ src/handler/authenticationHandler.js
@@ -2,66 +2,57 @@ const jwt = require('jsonwebtoken');
 const bcrypt = require('bcryptjs');
 const pool = require('../config/db');
 const refreshTokenHandler = require('./refreshTokenHandler');
+const ClientError = require('../exceptions/clientError');
+const InvariantError = require('../exceptions/invariantError');
+const NotFoundError = require('../exceptions/notFoundError');
+const AuthenticationError = require('../exceptions/authError');
 
 const signIn = async (req, res) => {
-  const { username, password } = req.body; // To get email and password input from body request
   try {
+    const { username, password } = req.body;
+
     const query = {
       text: 'SELECT id, username, password, role FROM users WHERE username=$1',
       values: [username],
     };
     const result = await pool.query(query);
-
     if (!result.rows.length) {
-      return res.status(404).send({
-        status: 'fail',
-        message: 'User not found!',
-      });
+      throw new NotFoundError(`${username} tidak ditemukan`);
     }
 
     const passwordIsValid = bcrypt.compareSync(
       password,
       result.rows[0].password,
     );
-
-    // If the password is not valid, send the error message
     if (!passwordIsValid) {
-      return res.status(401).send({
-        status: 'fail',
-        message: 'Invalid Password!',
-      });
-      // throw new AuthenticationError('Invalid Password!');
+      throw new AuthenticationError('Invalid Password!');
     }
 
-    // If password valid
-    try {
-      const token = jwt.sign({ user: result.rows[0].id }, process.env.SECRET, {
-        expiresIn: 120,
-      });
+    const token = jwt.sign({ user: result.rows[0].id }, process.env.SECRET, {
+      expiresIn: 120,
+    });
 
-      const refreshToken = await refreshTokenHandler.createToken(result.rows[0]);
+    const refreshToken = await refreshTokenHandler.createToken(result.rows[0]);
 
-      // If signIn Success, send these response
-      return res.status(200).send({
-        status: 'success',
-        message: 'Login Success',
-        userData: {
-          id: result.rows[0].id,
-          username: result.rows[0].username,
-          role: result.rows[0].role,
-        },
-        token,
-        refreshToken,
-      });
-
-      // if signIn failed, send these response
-    } catch (e) {
-      return res.status(500).send({
-        status: 'failed',
+    return res.status(200).send({
+      status: 'success',
+      message: 'Login Success',
+      userData: {
+        id_user: result.rows[0].id,
+        username: result.rows[0].username,
+        role: result.rows[0].role,
+      },
+      token,
+      refreshToken,
+    });
+  } catch (e) {
+    if (e instanceof ClientError) {
+      res.status(e.statusCode).send({
+        status: 'fail',
         message: e.message,
       });
     }
-  } catch (e) {
+
     return res.status(500).send({
       status: 'fail',
       message: 'Sorry there was a failure on our server.',
@@ -72,8 +63,8 @@ const signIn = async (req, res) => {
 const refreshToken = async (req, res) => {
   const { refreshToken: requestToken } = req.body;
 
-  if (requestToken == null) {
-    return res.status(403).json({ msg: 'Refresh Token is required' });
+  if (!requestToken) {
+    throw new InvariantError('Refresh token harus diisi');
   }
   try {
     const query = {
@@ -83,10 +74,7 @@ const refreshToken = async (req, res) => {
     const result = await pool.query(query);
 
     if (result.rows.length === 0) {
-      return res.status(401).send({
-        status: 'fail',
-        message: 'Your session was expired or Token is not valid! Please make new login request',
-      });
+      throw new AuthenticationError('Sesi Anda telah habis atau token tidak valid. Mohon lakukan login kembali');
     }
     if (refreshTokenHandler.verifyExpiration(result.rows[0])) {
       const queryDelete = {
@@ -94,10 +82,7 @@ const refreshToken = async (req, res) => {
         values: [result.rows[0].id],
       };
       await pool.query(queryDelete);
-      return res.status(401).send({
-        status: 'fail',
-        message: 'Your session was expired or Token is not valid! Please make new login request',
-      });
+      throw new AuthenticationError('Sesi Anda telah habis atau token tidak valid. Mohon lakukan login kembali');
     }
 
     const queryUser = {
@@ -116,6 +101,13 @@ const refreshToken = async (req, res) => {
       refreshToken: result.rows[0].token,
     });
   } catch (e) {
+    if (e instanceof ClientError) {
+      res.status(e.statusCode).send({
+        status: 'fail',
+        message: e.message,
+      });
+    }
+
     return res.status(500).send({
       status: 'fail',
       message: 'Sorry there was a failure on our server.',
@@ -127,6 +119,14 @@ const logOut = async (req, res) => {
   try {
     // eslint-disable-next-line no-shadow
     const { refreshToken } = req.body;
+    const qGetRefToken = {
+      text: 'SELECT token FROM refresh_token WHERE token = $1',
+      values: [refreshToken],
+    };
+    const resGetRef = await pool.query(qGetRefToken);
+    if (!resGetRef.rows.length) {
+      throw new NotFoundError('Token tidak ditemukan atau tidak valid');
+    }
 
     const queryDelete = {
       text: 'DELETE FROM refresh_token WHERE token=$1',
@@ -138,7 +138,17 @@ const logOut = async (req, res) => {
       message: 'Authentications has been removed',
     });
   } catch (e) {
-    return res.status(403).json(e.message);
+    if (e instanceof ClientError) {
+      res.status(e.statusCode).send({
+        status: 'fail',
+        message: e.message,
+      });
+    }
+
+    return res.status(500).send({
+      status: 'error',
+      messgae: e.message,
+    });
   }
 };
